refactor: extract shared linear mla support
